Remove webpack from core-http
Switched from webpack to rollup for browser testing.

To make this work as expected, I had to split some of the tests up into
separate files as the dynamic nature of checking isNode was pulling
in many unexpected things to the rollup bundle.
