+++ sdk/core/core-http/package.json
@@ -59,15 +59,14 @@
   },
   "scripts": {
     "audit": "node ../../../common/scripts/rush-audit.js && rimraf node_modules package-lock.json && npm i --package-lock-only 2>&1 && npm audit",
-    "build": "run-p build:lib",
-    "build:lib": "run-s build:tsc build:rollup",
+    "build": "run-s build:tsc build:rollup",
     "build:scripts": "tsc -p ./.scripts/",
     "build:test": "run-s build build:test-browser",
     "build:tsc": "tsc -p tsconfig.es.json",
-    "build:rollup": "rollup -c rollup.config.ts 2>&1",
-    "build:test-browser": "webpack --config webpack.testconfig.ts",
+    "build:rollup": "rollup -c 2>&1",
+    "build:test-browser": "rollup -c rollup.test.config.js 2>&1",
     "check-format": "prettier --list-different --config ../../.prettierrc.json \"lib/**/*.ts\" \"test/**/*.ts\" \"*.{js,json}\"",
-    "clean": "rimraf ./es ./dist",
+    "clean": "rimraf ./es ./dist ./dist-test",
     "execute:samples": "echo skipped",
     "format": "prettier --write --config ../../.prettierrc.json \"lib/**/*.ts\" \"test/**/*.ts\" \"*.{js,json}\"",
     "integration-test:browser": "echo skipped",
@@ -148,8 +147,6 @@
     "@types/sinon": "^7.0.13",
     "@types/tough-cookie": "^2.3.5",
     "@types/uuid": "^3.4.3",
-    "@types/webpack": "^4.4.13",
-    "@types/webpack-dev-middleware": "^2.0.2",
     "@types/xml2js": "^0.4.3",
     "@typescript-eslint/eslint-plugin": "^2.0.0",
     "@typescript-eslint/parser": "^2.0.0",
@@ -172,7 +169,6 @@
     "karma-rollup-preprocessor": "^7.0.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-typescript-es6-transform": "^4.0.0",
-    "karma-webpack": "^4.0.2",
     "mocha": "^6.2.2",
     "mocha-chrome": "^2.0.0",
     "mocha-junit-reporter": "^1.18.0",
@@ -194,9 +190,6 @@
     "ts-node": "^8.3.0",
     "typescript": "~3.6.4",
     "uglify-js": "^3.4.9",
-    "webpack": "^4.16.3",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-middleware": "^3.1.2",
     "xhr-mock": "^2.4.1"
   }
 }
