+++ sdk/core/core-http/rollup.config.js
@@ -1,0 +1,5 @@
+import * as base from "./rollup.base.config";
+
+const inputs = [base.nodeConfig(), base.browserConfig()];
+
+export default inputs;
