+++ sdk/core/core-http/rollup.test.config.js
@@ -1,0 +1,3 @@
+import * as base from "./rollup.base.config";
+
+export default [base.browserConfig(true)];
