+++ sdk/core/core-http/test/defaultHttpClientTests.browser.ts
@@ -1,0 +1,76 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import "chai/register-should";
+
+import { DefaultHttpClient } from "../lib/defaultHttpClient";
+import { WebResource, TransferProgressEvent } from "../lib/webResource";
+import { getHttpMock, HttpMockFacade } from "./mockHttp";
+
+describe("defaultHttpClient (browser)", function() {
+  let httpMock: HttpMockFacade;
+  beforeEach(() => {
+    httpMock = getHttpMock();
+    httpMock.setup();
+  });
+  afterEach(() => httpMock.teardown());
+  after(() => httpMock.teardown());
+
+  describe("should report upload and download progress", () => {
+    type Notified = { notified: boolean };
+    const listener = (operationStatus: Notified, ev: TransferProgressEvent) => {
+      operationStatus.notified = true;
+      if (typeof ProgressEvent !== "undefined") {
+        ev.should.not.be.instanceof(ProgressEvent);
+      }
+      ev.loadedBytes.should.be.a("Number");
+    };
+
+    it("for blob bodies", async function() {
+      let payload = new Blob([new Uint8Array(1024 * 1024)]);
+      const size = undefined;
+
+      httpMock.post("/bigfileupload", async (_url, _method, _body) => {
+        return {
+          status: 250,
+          body: payload,
+          headers: { "Content-Type": "text/javascript", "Content-length": size }
+        };
+      });
+
+      const upload: Notified = { notified: false };
+      const download: Notified = { notified: false };
+
+      const request = new WebResource(
+        "/bigfileupload",
+        "POST",
+        payload,
+        undefined,
+        undefined,
+        true,
+        undefined,
+        undefined,
+        0,
+        (ev) => listener(upload, ev),
+        (ev) => listener(download, ev)
+      );
+
+      const client = new DefaultHttpClient();
+      const response = await client.sendRequest(request);
+      response.status.should.equal(250);
+      if (response.blobBody) {
+        await response.blobBody;
+      } else if (typeof response.readableStreamBody === "function") {
+        const streamBody = (response.readableStreamBody as Function)();
+        streamBody.on("data", () => {});
+        await new Promise((resolve, reject) => {
+          streamBody.on("end", resolve);
+          streamBody.on("error", reject);
+        });
+      }
+
+      upload.notified.should.be.true;
+      download.notified.should.be.true;
+    });
+  });
+});
